ongoing fixes for dict hangs.
